Switch to the proper variable size.
This fixes a Win32 compiler warning.

+++ src/SerializeUtils.cpp
@@ -1,6 +1,6 @@
 // CommitMonitor - simple checker for new commits in svn repositories
 
-// Copyright (C) 2007, 2012-2013 - Stefan Kueng
+// Copyright (C) 2007, 2012-2014 - Stefan Kueng
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -81,7 +81,7 @@ bool CSerializeUtils::SaveString(FILE * hFile, std::string str)
     SerializeTypes type = SerializeType_String;
     if (fwrite(&type, sizeof(type), 1, hFile))
     {
-        int length = (int)str.size();
+        size_t length = str.size();
         if (fwrite(&length, sizeof(length), 1, hFile))
         {
             if (fwrite(str.c_str(), sizeof(char), length, hFile)>=length)
